src/aiserver/requirements.txt
# Web Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Face Recognition System Dependencies
numpy
opencv-python
insightface
openai
watchdog
onnxruntime